moved btn-radio directive to vis editor
